--- src/components/grid/userPhotoGallery/WishListGrid.tsx
@@ -1,33 +1,0 @@
-import gridStyle from "./UserPhotoGallery.module.css";
-import getWishList from "api/wishlist/getWishList";
-import WishListThumbnailCard from "components/card/wishList/WishListThumbnailCard";
-import { IWishListDTO } from "interface/wish/IWishList";
-import { useEffect, useState } from "react";
-
-const WishListGrid = () => {
-  const [wishList, setWishList] = useState<IWishListDTO[]>([]);
-  useEffect(
-    () => {
-      const fetchWishList = async () => {
-        const data = await getWishList();
-        if (data) setWishList(data);
-        console.log(data);
-      };
-      fetchWishList();
-    },
-    // eslint-disable-next-line
-    []
-  );
-  return (
-    <div className={gridStyle.gridGallery}>
-      {wishList.map((wishList) => (
-        <WishListThumbnailCard
-          key={wishList.wishListId}
-          wishListId={wishList.wishListId}
-          name={wishList.name}
-        />
-      ))}
-    </div>
-  );
-};
-export default WishListGrid;
